\name{linmod-package}
\alias{linmod-package}
\docType{package}
\title{Linear and Generalized Linear Moddels}
\description{
  TODO
}
\details{
\tabular{ll}{
  Package: \tab linmod\cr
  Type: \tab Package\cr
  License: \tab MPL 2.0\cr
}
}
\author{
  Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
