% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex_function.R
\name{cindex}
\alias{cindex}
\title{Condition Index Function}
\usage{
cindex(l)
}
\arguments{
\item{l}{a vector of eigen values}
}
\description{
This function creates generates the condition index from a vector of eigen values
}
\examples{
cindex_function()
}
\keyword{cindex}
