% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createpower_function.R
\name{createpower}
\alias{createpower}
\title{Generate Normal Power Function}
\usage{
createpower(n, a)
}
\arguments{
\item{n}{sample size}

\item{a}{alpha value}
}
\description{
This function creates a power function under a normal distribution
}
\examples{
createpower_function()
}
\keyword{createpower}
