% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvi_function.R
\name{eigenvi}
\alias{eigenvi}
\title{Eigen Value and Condition Index Function}
\usage{
eigenvi(mat)
}
\arguments{
\item{mat}{a matrix}
}
\description{
This function creates a table with eigen values and condition index from a matrix
}
\examples{
eigenvi_function()
}
\keyword{eigenvi}
