\name{glm_fit}
\alias{glm_fit}

\title{Generalized Linear Model Fitter}

\description{
  TODO
}
\usage{
}
\arguments{
}
\details{
  TODO
}
\value{
  A list containing the elements:
\tabular{l}{
  coefficients \cr
  residuals \cr
  fitted.values \cr
  effects \cr
  R \cr
  rank \cr
  qr \cr
  family \cr
  linear.predictors \cr
  deviance \cr
  aic \cr
  null.deviance \cr
  iter \cr
  weights \cr
  prior.weights \cr
  df.residual \cr
  df.null \cr
  y \cr
  converged \cr
  boundary \cr
}
  where \code{qr} is a list consisting of the elements:
\tabular{l}{
  qr \cr
  qraux \cr
  pivot \cr
  tol \cr
  rank \cr
}
}
\examples{
\dontrun{
library(linmod)

TODO
}
}
