% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Linear Model}
\usage{
linear_model(Y, X)
}
\arguments{
\item{Y}{input outcome}

\item{X}{input predictor(s)}
}
\value{
estimated effects of X on Y through OLS
}
\description{
Calculates coefficient estimates, standard errors, t-statistics, and p-values,
as well as R squared, adjusted R squared, F-Statistic and p-value for model.
}
