% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05e_Uebung.R
\name{linmod.default}
\alias{linmod.default}
\alias{print.linmod}
\alias{coef.linmod}
\title{Internal methods}
\usage{
\method{linmod}{default}(x, y, ...)

\method{print}{linmod}(x, ...)

\method{coef}{linmod}(object, ...)
}
\arguments{
\item{x}{object of class linmod}

\item{y}{numeric vector. Dependent variable.}

\item{\dots}{further objects to be passed to methods}

\item{object}{object of class linmod for coef-method. Argument titled `object`
instead of `x` in order to be consistent with stats::coef.
Otherwise CHECK --as-cran throws a warning.}
}
\description{
preexists in base; therefor just add class method
}
