\name{linmod-package}
\alias{linmod-package}
\docType{package}
\title{
What the package does (short line)

}
\description{
More about what it does (maybe more than one line)}
\details{
\tabular{ll}{
Package: \tab linmod\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-07\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it
}
\references{
dfh}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
fdfgh}
\examples{}
