\name{linmod}
\alias{linmod}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
sadf}
\description{
sadf}
\usage{
linmod(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
asf}
  \item{\dots}{
sadf}
}
\details{
sadf}
\value{
sadf%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
asf}
\author{
af}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
sadf}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x, ...) 
UseMethod("linmod")
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
