% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunc_function.R
\name{plotfunc}
\alias{plotfunc}
\title{Plot Function}
\usage{
plotfunc(f, lower, upper)
}
\arguments{
\item{f}{function}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\description{
Plots a function
}
\examples{
plotfunc_function()
}
\keyword{plotfunc}
