% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmult_function.R
\name{plotmult}
\alias{plotmult}
\title{Plot Mult Function}
\usage{
plotmult(v, a, lower = 0, upper = 1)
}
\arguments{
\item{v}{vector of values}

\item{a}{alpha value}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\description{
Plots several functions
}
\examples{
plotmult_function()
}
\keyword{plotfunc}
