% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpower_function.R
\name{plotpower}
\alias{plotpower}
\title{Plot Normal Power Function}
\usage{
plotpower(n, a)
}
\arguments{
\item{n}{sample size}

\item{a}{alpha value}
}
\description{
This function creates and plots a normal power function requires ggplot
}
\examples{
plotpower_function()
}
\keyword{plotpower}
