% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerfun_function.R
\name{powerfun}
\alias{powerfun}
\title{Generate Normal Power Function}
\usage{
powerfun(n, a)
}
\arguments{
\item{n}{sample size}

\item{a}{alpha value}
}
\description{
This function creates a power function under a normal distribution
}
\examples{
powerfun_function()
}
\keyword{powerfun}
