% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerfun2_function.R
\name{powerfun2}
\alias{powerfun2}
\title{Generate two-tailed Normal Power Function}
\usage{
powerfun2(n, a)
}
\arguments{
\item{n}{sample size}

\item{a}{alpha value}
}
\description{
This function creates a two-tailed power function under a normal distribution
}
\examples{
powerfun2_function()
}
\keyword{powerfun}
