\name{print.linmod}
\alias{print.linmod}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
safd}
\description{
saf}
\usage{
print.linmod(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
dgdfsg}
  \item{\dots}{
dsf}
}
\details{
dsf}
\value{
dsfg%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
dfsg}
\author{
dfsg}
\note{
dsfg}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
dsfg}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x, ...) 
{
    cat("Call:\n")
    print(x$call)
    cat("\nCoefficients:\n")
    print(x$coefficients)
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
