\name{print.myvector}
\alias{print.myvector}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
sfa}
\description{
sdf}
\usage{
print.myvector(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
sadf}
  \item{\dots}{
saf}
}
\details{
saf}
\value{
saf%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
sadf}
\author{
saf}
\note{
saf}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
sdf}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x, ...) 
{
    cat("This is my vector:\n")
    cat(paste(x[1:5]), "...\n")
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
