\name{print.summary.linmod}
\alias{print.summary.linmod}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
asfd}
\description{
asdf}
\usage{
print.summary.linmod(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
sadf}
  \item{\dots}{
saf}
}
\details{
sf}
\value{
sdaf%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
sadf}
\author{
saf}
\note{
fsa}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
asf}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x, ...) 
{
    cat("Call:\n")
    print(x$call)
    cat("\n")
    printCoefmat(x$coefficients, P.value = TRUE, has.Pvalue = TRUE)
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
