% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\docType{class}
\name{ridgereg}
\alias{ridgereg}
\title{RC-class ridge regression model}
\source{
\url{https://en.wikipedia.org/wiki/Ridge_regression}
}
\arguments{
\item{formula}{A formula denoted as 'Y ~ X' that formalizes the relationship to be linearly calculated.
The independent variables on the right side of the tilde sign can include multiple,
separated by '+'.}

\item{data}{A data frame that contains the variables which are to be modelled.}

\item{lambda}{Hyperparameter to tune the model coefficients}
}
\value{
An RC-class object of the created ridge regression model on which
 analyze can be made.
}
\description{
An object of RC-class that is used to create a ridge regression model
for a given formula and dataset. The created model can then be used to access varying functions for analyzing.
Functions available are: print, pred and coef.
}
\section{Fields}{

\describe{
\item{\code{formula}}{A ridge regression-formula.}

\item{\code{data}}{A data frame with variables for the model.}

\item{\code{Y}}{A vector for the response variable.}

\item{\code{X}}{A matrix of the x variables.}

\item{\code{beta_hat}}{A matrix with the coefficients.}

\item{\code{y_hat}}{A vector with predicted values.}

\item{\code{Xnrom}}{A matrix with scaled variable values}

\item{\code{df_name}}{A character vector with the name of the data frame.}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{The coefficients as a named vector}

\item{\code{pred(newdata = data)}}{Return the predicted Y values of training data or new data}

\item{\code{print()}}{Printing the formula and the coefficients}
}}

