% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledsscp_function.R
\name{scaledsscp}
\alias{scaledsscp}
\title{Scaled SSCP Matrix Function}
\usage{
scaledsscp(x)
}
\arguments{
\item{x}{design matrix}
}
\description{
This function creates a scaled sscp matrix from a design matrix.
}
\examples{
scaledsscp_function()
}
\keyword{scaledsscp}
