% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_function.R
\name{tf}
\alias{tf}
\title{TF factor Function}
\usage{
tf(x)
}
\arguments{
\item{x}{logical variable}
}
\description{
This function creates a factor variable from a logical
}
\examples{
tf_function()
}
\keyword{tf}
