% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{dplyr plot}
\usage{
visualize_airport_delays()
}
\description{
A function which uses dplyr on the airports dataset from the nycflights13 package 
and manipulates it and returns a plot.
}
