\name{x}
\alias{x}
\docType{data}
\title{
sdfg}
\description{
dsg}
\usage{data(x)}
\format{
  The format is:
 num [1:144, 1:2] 1 1 1 1 1 1 1 1 1 1 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:2] "Const" "Bwt"
}
\details{
dg}
\source{
sdg}
\references{
dsg}
\examples{
data(x)
## maybe str(x) ; plot(x) ...
}
\keyword{datasets}
