% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glm_fit.r
\name{glm_fit}
\alias{glm_fit}
\title{Generalized Linear Model Fitter}
\usage{
glm_fit(x, y, offset = rep(0, nobs), family = gaussian(),
  control = list(), intercept = TRUE, ..., stoprule = "deviance")
}
\value{
A list containing the elements:
\tabular{l}{
    coefficients \cr
    residuals \cr
    fitted.values \cr
    effects \cr
    R \cr
    rank \cr
    qr \cr
    family \cr
    linear.predictors \cr
    deviance \cr
    aic \cr
    null.deviance \cr
    iter \cr
    weights \cr
    prior.weights \cr
    df.residual \cr
    df.null \cr
    y \cr
    converged \cr
    boundary \cr
}
where \code{qr} is a list consisting of the elements:
\tabular{l}{
    qr \cr
    qraux \cr
    pivot \cr
    tol \cr
    rank \cr
}
}
\description{
TODO
}
\details{
TODO
}
\examples{
\dontrun{
library(linmod)

TODO
}
}

