\name{linmod-package}
\alias{linmod-package}
\alias{linmod}
\docType{package}
\title{
\packageTitle{linmod}
}
\description{
\packageDescription{linmod}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{linmod}
\packageIndices{linmod}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{linmod}

Maintainer: \packageMaintainer{linmod}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
