% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linmod-package.R
\docType{package}
\name{linmod-package}
\alias{linmod}
\alias{linmod-package}
\title{linmod: Create a linear model from a given formula}
\description{
This package is used to create a linear or a ridge model by supplying the formula and dataset. The model is calculated using least squares and returns an object of class RC. Within the object you can call the following methods to analyze your model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Johhed15/Lab4_R}
}

}
\author{
\strong{Maintainer}: Johannes Hedström \email{johed883@student.liu.se}

Authors:
\itemize{
  \item Mikael Montén \email{mikmo937@student.liu.se}
  \item Siddhesh Sreedar \email{sidsr770@student.liu.se}
}

}
\keyword{internal}
