% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linmod-package.R
\docType{package}
\name{linmod-package}
\alias{linmod-package}
\title{Linear and Generalized Linear Moddels}
\description{
The linmod package contains wrappers of the linmod library, a collection of
utilities for fitting linear and generalized linear models.  These use
modified versions of LAPACK routines (compared to R's use of LINPACK
routines).
}
\details{
\tabular{ll}{
   Package: \tab linmod \cr
   Type: \tab Package \cr
   License: \tab MPL 2.0\cr
}
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}

