% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linmod.R
\name{linmod}
\alias{linmod}
\alias{linmod.default}
\alias{linmod.formula}
\alias{predict.linmod}
\alias{print.linmod}
\alias{print.summary.linmod}
\alias{summary.linmod}
\title{Linear Regression}
\usage{
linmod(x, ...)

\method{linmod}{default}(x, y, ...)

\method{linmod}{formula}(formula, data = list(), ...)

\method{print}{linmod}(x, ...)

\method{summary}{linmod}(object, ...)

\method{print}{summary.linmod}(x, ...)
}
\arguments{
\item{x}{a numeric design matrix for the model.}

\item{y}{a numeric vector of responses.}

\item{formula}{a symbolic description of the model to be fit.}

\item{data}{an optional data frame containing the variables in the model.}

\item{object}{an object of class \code{"linmod"}, i.e., a fitted model.}

\item{\dots}{not used.}
}
\value{
An object of class \code{logreg}, basically a list including
elements \item{coefficients}{ a named vector of coefficients } \item{vcov}{
covariance matrix of coefficients } \item{fitted.values}{ fitted values }
\item{residuals}{ residuals }
}
\description{
Fit a linear regression model.
}
\examples{

data(cats, package="MASS")
mod1 <- linmod(Hwt~Bwt*Sex, data=cats)
mod1
summary(mod1)

}
\author{
Friedrich Leisch
}
\keyword{regression}

