% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05e_Uebung.R
\name{linmodEst}
\alias{linmodEst}
\alias{linmod}
\title{main estimation function}
\usage{
linmodEst(x, y)

linmod(x, ...)
}
\arguments{
\item{x}{numeric perdictor matrix}

\item{y}{numeric vector. Dependent variable.}

\item{\dots}{further objects to be passed to methods}
}
\value{
List of 4 elements
 \itemize{
    \item coefficients coefficients of parameter estimation
    \item vcov variance covaraiance matrix
    \item sigma residual standard deviation
    \item df degrees of freedom
 }
}
\description{
Takes in a matrix x of predictors and computes linear model
of y on x.
}
