% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\docType{class}
\name{linreg}
\alias{linreg}
\title{RC-class object to create linear model}
\source{
\url{https://en.wikipedia.org/wiki/Ordinary_least_squares}
}
\arguments{
\item{formula}{A formula denoted as 'Y ~ X' that formalizes the relationship to be linearly calculated.
The independent variables on the right side of the tilde sign can include multiple, separated by '+'.}

\item{data}{A data frame that contains the variables which are to be modelled.}
}
\value{
An RC-class object of the created linear model on which analyze can be made.
}
\description{
An object of RC-class that is used to create a linear model from given formula.
The formula should have variables that stems from a given dataset and formalized as 'Y ~ X'.
The created model can then be used to access varying functions for analyzing.
Functions available are: print, plot, resid, pred, coef, summary.
}
\section{Methods}{

\describe{
\item{\code{coef()}}{The coefficients as a named vector}

\item{\code{plot()}}{creating a dataframe to create plots}

\item{\code{pred()}}{Return the predicted Y values}

\item{\code{print()}}{Printing the formula and the coefficients}

\item{\code{resid()}}{Return the residuals as a vector}

\item{\code{summary()}}{The summary table which looks like the lm summary()}
}}

