% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpower_function.R
\name{plotpower}
\alias{plotpower}
\title{Plot Normal Power Function}
\usage{
plotpower(n, a, lower = 0, upper = 1, twotail = F)
}
\arguments{
\item{n}{sample size}

\item{a}{alpha value}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{twotail}{twotail T/F}
}
\description{
This function creates and plots a normal power function requires ggplot
}
\examples{
plotpower_function()
}
\keyword{plotpower}
