% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linmod.R
\name{summary.linmod}
\alias{summary.linmod}
\title{Summary Linear Regression}
\usage{
\method{summary}{linmod}(object, ...)
}
\arguments{
\item{object}{= object}

\item{...}{= parameter}
}
\description{
Summary Linear Regression
}
